# Optimize default CPU path of Convolution with MKLDNN