@ngimel 2020-12-06
1x1 convolutions directly use gemm calls now in pytorch, w/o any overhead. What is mkldnn doing that's better than mkl? Should you guys fix mkl for those cases?