@pinzhenx 2020-12-06
@ngimel For 1x1 conv, we also directly call the sgemm rewritten by MKLDNN. So it's reasonable that we have a small margin over MKL-based THNN conv