@CaoZhongZ 2020-12-06
> 1x1 convolutions directly use gemm calls now in pytorch, w/o any overhead. What is mkldnn doing that's better than mkl? Should you guys fix mkl for those cases?

We removed unnecessary transforms in 1x1 convolution and release the performance as it was supposed to be. The difference between MKL SGEMM and oneDNN should be trivial, the only difference is we tested and targeted oneDNN for DNN usage cases. We could look into why oneDNN is faster than MKL call in THNN.