@gottbrath 2021-01-27
Comment from Jan 27th discussion: 
* Please benchmark & test on non AVX512 use cases (i.e. AVX2)
* please explicitly state which version of MKLDNN is required.
* Commit benchmark scripts as part of this PR. (@VitalyFedyunin  to follow up with instructions)